Move navigation toolbar class into separate file
This is so that we can use it in other parts of the code.
