+++ hexrd/ui/image_tab_widget.py
@@ -1,13 +1,12 @@
 from PySide2.QtCore import Signal, Slot, Qt
 from PySide2.QtWidgets import QMessageBox, QTabWidget, QHBoxLayout
 
-from matplotlib.backends.backend_qt5agg import NavigationToolbar2QT
-
 import numpy as np
 
 from hexrd.ui.hexrd_config import HexrdConfig
 from hexrd.ui.image_canvas import ImageCanvas
 from hexrd.ui.image_series_toolbar import ImageSeriesToolbar
+from hexrd.ui.navigation_toolbar import NavigationToolbar
 from hexrd.ui import utils
 
 
@@ -124,8 +123,7 @@ class ImageTabWidget(QTabWidget):
         while len(self.toolbars) != len(self.image_canvases):
             # The new one to add
             idx = len(self.toolbars)
-            tb = CustomNavigationToolbar2QT(self.image_canvases[idx], parent,
-                                            False)
+            tb = NavigationToolbar(self.image_canvases[idx], parent, False)
             # Current detector
             name = self.image_names[idx]
             sb = ImageSeriesToolbar(name, self)
@@ -281,26 +279,6 @@ class ImageTabWidget(QTabWidget):
 
     def polar_show_snip1d(self):
         self.image_canvases[0].polar_show_snip1d()
-
-
-class CustomNavigationToolbar2QT(NavigationToolbar2QT):
-
-    def __init__(self, canvas, parent, coordinates=True):
-        # Remove these buttons from the navigation toolbar
-        nav_toolbar_blacklist = [
-            'Subplots'
-        ]
-        old_toolitems = NavigationToolbar2QT.toolitems
-        NavigationToolbar2QT.toolitems = [
-            x for x in NavigationToolbar2QT.toolitems
-            if x[0] not in nav_toolbar_blacklist
-        ]
-
-        super(CustomNavigationToolbar2QT, self).__init__(canvas, parent, coordinates)
-
-        # Restore the global navigation tool items for other parts of
-        # the program to use them.
-        NavigationToolbar2QT.toolitems = old_toolitems
 
 
 if __name__ == '__main__':
